アプリの処理を別ファイルに分ける(useApp))
